examples/s2s/requirements.txt
soundfile
openai-whisper==20231117
transformers==4.51.0
librosa
numpy==1.26.4
HyperPyYAML==1.2.2
modelscope==1.15.0
onnxruntime-gpu==1.18.0; sys_platform == 'linux'
onnxruntime==1.18.0; sys_platform == 'darwin' or sys_platform == 'windows'
WeTextProcessing==1.0.3
conformer==0.3.2
diffusers>=0.29.0
gradio==5.3.0
grpcio==1.57.0
grpcio-tools==1.57.0
huggingface_hub>=0.30.0
inflect==7.3.1
matplotlib==3.7.5
lightning==2.2.4
torchaudio==2.3.0
wget==3.2
num2words==0.5.14
torchvision==0.18.0
deepspeed==0.16.7
peft==0.15.2